Flatten graph Occurence - move it to a separate source
