+++ engine/runtime-compiler/src/main/java/org/enso/compiler/pass/analyse/PassPersistance.java
@@ -3,6 +3,7 @@ package org.enso.compiler.pass.analyse;
 import java.io.IOException;
 import org.enso.compiler.pass.analyse.alias.AliasMetadata;
 import org.enso.compiler.pass.analyse.alias.graph.Graph;
+import org.enso.compiler.pass.analyse.alias.graph.Occurrence;
 import org.enso.compiler.pass.analyse.types.TypeInference;
 import org.enso.compiler.pass.resolve.DocumentationComments;
 import org.enso.compiler.pass.resolve.DocumentationComments$;
@@ -60,14 +61,8 @@ import scala.Tuple2$;
 @Persistable(clazz = AliasMetadata.Occurrence.class, id = 1261, allowInlining = false)
 @Persistable(clazz = AliasMetadata.RootScope.class, id = 1262, allowInlining = false)
 @Persistable(clazz = AliasMetadata.ChildScope.class, id = 1263, allowInlining = false)
-@Persistable(
-    clazz = org.enso.compiler.pass.analyse.alias.graph.Graph$Occurrence$Use.class,
-    id = 1264,
-    allowInlining = false)
-@Persistable(
-    clazz = org.enso.compiler.pass.analyse.alias.graph.Graph$Occurrence$Def.class,
-    id = 1265,
-    allowInlining = false)
+@Persistable(clazz = Occurrence.Use.class, id = 1264, allowInlining = false)
+@Persistable(clazz = Occurrence.Def.class, id = 1265, allowInlining = false)
 @Persistable(clazz = Graph.Link.class, id = 1266, allowInlining = false)
 @Persistable(clazz = TypeInference.class, id = 1280)
 public final class PassPersistance {
@@ -125,7 +120,7 @@ public final class PassPersistance {
     @SuppressWarnings("unchecked")
     protected Graph.Scope readObject(Input in) throws IOException {
       var childScopes = in.readInline(scala.collection.immutable.List.class);
-      var occurrencesValues = (scala.collection.immutable.Set<Graph.Occurrence>) in.readObject();
+      var occurrencesValues = (scala.collection.immutable.Set<Occurrence>) in.readObject();
       var occurrences = occurrencesValues.map(v -> Tuple2$.MODULE$.apply(v.id(), v)).toMap(null);
       var allDefinitions = in.readInline(scala.collection.immutable.List.class);
       var parent = new Graph.Scope(childScopes, occurrences, allDefinitions);
